+++ src/bindings/PyDP/algorithms/partition_selection_strategies.cpp
@@ -7,7 +7,6 @@
 
 #include "algorithms/partition-selection.h"
 
-
 namespace py = pybind11;
 namespace dp = differential_privacy;
 
@@ -28,8 +27,7 @@ PyPartitionSelectionStrategy CreateTruncatedGeometricPartitionStrategy(
 }
 
 PyPartitionSelectionStrategy CreateLaplacePartitionStrategy(
-    double epsilon, double delta, int max_partitions_contributed
-    ) {
+    double epsilon, double delta, int max_partitions_contributed) {
   dp::LaplacePartitionSelection::Builder builder;
   builder.SetEpsilon(epsilon);
   builder.SetDelta(delta);
@@ -43,8 +41,7 @@ PyPartitionSelectionStrategy CreateLaplacePartitionStrategy(
 }
 
 PyPartitionSelectionStrategy CreateGaussianPartitionStrategy(
-    double epsilon, double delta, int max_partitions_contributed
-    ) {
+    double epsilon, double delta, int max_partitions_contributed) {
   dp::GaussianPartitionSelection::Builder builder;
   builder.SetEpsilon(epsilon);
   builder.SetDelta(delta);
@@ -58,22 +55,22 @@ PyPartitionSelectionStrategy CreateGaussianPartitionStrategy(
 }
 
 void init_algorithms_partition_selection_strategies(py::module& m) {
-  auto pyClass = py::class_<dp::PartitionSelectionStrategy, PyPartitionSelectionStrategy>(m, "PartitionSelectionStrategy",
-      R"pbdoc(
+  auto pyClass =
+      py::class_<dp::PartitionSelectionStrategy, PyPartitionSelectionStrategy>(
+          m, "PartitionSelectionStrategy",
+          R"pbdoc(
         Base class for all (Ɛ, 𝛿)-differenially private partition selection strategies.
       )pbdoc");
-  pyClass.def("should_keep",
-            &dp::PartitionSelectionStrategy::ShouldKeep,
-            py::arg("num_users"),
-            R"pbdoc(
+  pyClass
+      .def("should_keep", &dp::PartitionSelectionStrategy::ShouldKeep,
+           py::arg("num_users"),
+           R"pbdoc(
               Decides whether or not to keep a partition with `num_users` based on differential privacy parameters and strategy.
             )pbdoc")
       .attr("__module__") = "_partition_selection";
 
   m.def("create_truncted_geometric_partition_strategy",
         &CreateTruncatedGeometricPartitionStrategy);
-  m.def("create_laplace_partition_strategy", 
-        &CreateLaplacePartitionStrategy);
-  m.def("create_gaussian_partition_strategy", 
-        &CreateGaussianPartitionStrategy);
+  m.def("create_laplace_partition_strategy", &CreateLaplacePartitionStrategy);
+  m.def("create_gaussian_partition_strategy", &CreateGaussianPartitionStrategy);
 }
