+++ src/pydp/__init__.py
@@ -6,7 +6,5 @@ from pydp import algorithms
 from pydp import distributions
 from pydp import util
 from pydp import ml
-from pydp import partition_selection
 
 __version__ = "1.1.0"
-
