+++ src/pydp/algorithms/__init__.py
@@ -1,4 +1,9 @@
 # pydp relative
 from . import laplacian
+from .partition_selection import PartitionSelectionStrategy, create_partition_strategy
 
-__all__ = ["laplacian"]
+__all__ = [
+    "laplacian",
+    "PartitionSelectionStrategy",
+    "create_partition_strategy",
+]
